Backend/python/requirements.txt
psycopg[binary]==3.1.18
python-dotenv==1.0.1